Completed Example 11
Transtioned to all getter methods: getVal(), getNewVal(), updated(), using templates for all floats.

Finalized templates for getVal and getNewVal, including making <int>  default so it does not have to be set for most getVal() and getNewVal() calls.  Works for booleans as well.  TO DO:  Re-work and check ALL prior examples to ensure they use getVal, etc., and DON'T access value, newValue, isUpdated, directly.

+++ src/extras/PwmPin.cpp
@@ -33,7 +33,7 @@ void PwmPin::set(uint8_t channel, uint8_t level){
 
 ///////////////////
 
-void PwmPin::HSVtoRGB(double h, double s, double v, double *r, double *g, double *b ){
+void PwmPin::HSVtoRGB(float h, float s, float v, float *r, float *g, float *b ){
 
   // The algorithm below was provided on the web at https://www.cs.rit.edu/~ncs/color/t_convert.html
   // h = [0,360]
@@ -41,7 +41,7 @@ void PwmPin::HSVtoRGB(double h, double s, double v, double *r, double *g, double
   // v = [0,1]
 
   int i;
-  double f, p, q, t;
+  float f, p, q, t;
   
   if( s == 0 ){
     *r = *g = *b = v;
